iterate using char instead of int
In my opinion, this improves readability by directly iterating over the characters in the alphabet.
Furthermore, it improves performance by saving the `+` in the line that was previously `str[i] = 'a' + c`

Overall, it's a minor improvement. Either way, I thought you might like it so decided to suggest it.