practice/primitives: add improved version of parity from O(n) to O(k)
Instead of going through every bit of the input, we only iterate over
the set bits. This is achieved by droppong off the lowest set bit of the
input after each loop.

The operation x & (x-1) drops off the lowest set bit of x.

To flip the bit, the ^ (XOR) operation is used